+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-Detailed CHANGELOG 
+Detailed CHANGELOG
 ==================
 
 olefile.py
 ----------
+* 2017-11-05 v0.45 PL: - fixed issue #31, set all attributes in OleFileIO.\_\_init__
+
 (TODO: reverse order to have latest changes on top, use MD formatting)
 ```
 History from PIL 1.1.6: